+++ keras/backend/mxnet_backend.py
@@ -3940,3 +3940,247 @@ def _convnd(x, kernel, strides, filter_dilation, name=None, padding_mode='valid'
     # Handle original Data Format
     result = _postprocess_convnd_output(KerasSymbol(conv), data_format)
     return result
+
+
+def get_model():
+    """Prepares Model class that can be used for training a Keras model with MXNet backend.
+    Inherits and extends keras.engine.Model class.
+
+    # Returns
+        MXNetModel reference
+    """
+    import importlib
+    engine = importlib.import_module('keras.engine.training')
+
+    class MXNetModel(engine.Model):
+        def __init__(self, inputs, outputs, name=None, context=None, kvstore='device', **kwargs):
+            super(MXNetModel, self).__init__(inputs, outputs, name)
+            self._num_data = len(self.inputs)
+            self._num_label = len(self.outputs) + len(self.output_names)
+
+            # MXNet context
+            def get_mxnet_context(context_name):
+                if context_name.startswith('cpu'):
+                    return mx.cpu()
+                elif context_name.startswith('gpu('):
+                    index = int(context_name[4:-1])
+                    return mx.gpu(index)
+                elif context_name.startswith('gpu'):
+                    index = int(context_name[3:])
+                    return mx.gpu(index)
+
+            if context is None:
+                self._context = [mx.current_context()]
+            else:
+                if isinstance(context, str):
+                    self._context = [context]
+                self._context = [get_mxnet_context(context_name) for context_name in context]
+
+            self._kvstore = kvstore
+
+            self._data_names = None
+            self._label_names = None
+            self._ntrain = None
+            self._train_mxnet_symbol = None
+            self._train_updates = None
+            self._ntest = None
+            self._test_mxnet_symbol = None
+            self._test_updates = None
+            self._npred = None
+            self._pred_mxnet_symbol = None
+            self._arg_names = None
+            self._aux_names = None
+            self._fixed_weights = None
+            self._args = None
+            self._auxs = None
+            self._weights_dirty = None
+            self._module = None
+
+        def compile(self, optimizer, loss, metrics=None, loss_weights=None,
+                    sample_weight_mode=None, **kwargs):
+            super(MXNetModel, self).compile(
+                optimizer, loss, metrics, loss_weights,
+                sample_weight_mode, **kwargs)
+
+            # set the data and label
+            self._data_names = [x.name for x in self.inputs]
+            self._label_names = [x.name for x in self.targets + self.sample_weights]
+
+            # set for training
+            old = learning_phase()
+            set_learning_phase(1)
+            self._ntrain = len(self.metrics_tensors) + 1
+            train_updates = [stop_gradient(x[1]) for x in self.updates]
+            train_keras_symbol = group(
+                [make_loss(self.total_loss)] + [stop_gradient(x)
+                                                for x in self.metrics_tensors] + train_updates
+            )
+            bind_values = dfs_get_bind_values(train_keras_symbol)
+            self._train_mxnet_symbol = train_keras_symbol.symbol
+            symbol_name_map = {i.name: j.name for (_, i), j in zip(self.updates, train_updates)}
+            self._train_updates = {dst.name: symbol_name_map[src.name] for dst, src in self.updates}
+
+            # set for testing
+            set_learning_phase(0)
+            self._ntest = len(self.metrics_tensors) + 1
+            state_updates = [x[1] for x in self.state_updates]
+            test_keras_symbol = group(
+                [self.total_loss] +
+                [stop_gradient(x) for x in self.metrics_tensors] +
+                state_updates
+            )
+            bind_values.update(dfs_get_bind_values(test_keras_symbol))
+            self._test_mxnet_symbol = test_keras_symbol.symbol
+
+            # set for prediction
+            self._npred = len(self.outputs)
+            pred_keras_symbol = group(
+                self.outputs +
+                [symbol for symbol in state_updates if symbol not in self.outputs]
+            )
+            bind_values.update(dfs_get_bind_values(pred_keras_symbol))
+            self._pred_mxnet_symbol = pred_keras_symbol.symbol
+            self._test_updates = {dst.name: src.name for dst, src in self.state_updates}
+            set_learning_phase(old)
+
+            # set the args and auxs
+            inputs_name_set = set(self._data_names + self._label_names)
+            self._arg_names = set([x for x in self._train_mxnet_symbol.list_arguments()
+                                   if x not in inputs_name_set])
+            self._aux_names = set(self._train_mxnet_symbol.list_auxiliary_states())
+
+            trainable_weights = set([x.name for x in self.trainable_weights])
+            self._fixed_weights = [x for x in self._arg_names if x not in trainable_weights]
+            self._args = {x: bind_values[x] for x in self._arg_names}
+            self._auxs = {x: bind_values[x] for x in self._aux_names}
+            self._weights_dirty = False
+
+            # set the module
+            def sym_gen(phase):
+                if phase == 'train':
+                    return self._train_mxnet_symbol, self._data_names, self._label_names
+                elif phase == 'test':
+                    return self._test_mxnet_symbol, self._data_names, self._label_names
+                else:
+                    return self._pred_mxnet_symbol, self._data_names, None
+
+            self._module = mx.mod.BucketingModule(
+                sym_gen=sym_gen,
+                default_bucket_key='pred',
+                context=self._context,
+                fixed_param_names=self._fixed_weights)
+            set_model(self)
+
+        def _adjust_module(self, inputs, phase):
+            if not hasattr(self, '_module'):
+                raise RuntimeError('You must compile your model before using it.')
+            if self._num_data + self._num_label == len(inputs) - 1:
+                inputs = inputs[:-1]
+            elif self._num_data == len(inputs) - 1:
+                inputs = inputs[:-1]
+            assert self._num_data == len(inputs) or self._num_data + self._num_label == len(inputs)
+            data = [mx.nd.array(x, dtype=s.dtype)
+                    for (s, x) in zip(self.inputs, inputs[:self._num_data])]
+            data_shapes = [mx.io.DataDesc(s.name, arr.shape, dtype=s.dtype)
+                           for (s, arr) in zip(self.inputs, data)]
+            if self._num_data < len(inputs):
+                label = [mx.nd.array(x, dtype=s.dtype)
+                         for (s, x) in zip(self.targets + self.sample_weights,
+                                           inputs[self._num_data:])]
+                label_shapes = [mx.io.DataDesc(s.name, arr.shape, dtype=s.dtype)
+                                for (s, arr) in zip(self.targets + self.sample_weights, label)]
+            else:
+                label = None
+                label_shapes = None
+
+            if not self._module.binded:
+                self._module.bind(data_shapes=data_shapes, label_shapes=None, for_training=True)
+                self._set_weights()
+                self._module.init_optimizer(kvstore=self._kvstore, optimizer=self.optimizer)
+            self._module.switch_bucket(phase, data_shapes, label_shapes)
+
+            # adjust module data shape
+            if inputs[0].shape[0] != self._module._curr_module._exec_group.batch_size:
+                self._module._curr_module.reshape(data_shapes, label_shapes)
+                assert inputs[0].shape[0] == self._module._curr_module._exec_group.batch_size, \
+                    "Reshape failed"
+
+            return data, label, phase, data_shapes, label_shapes
+
+        def _sync_weights(self):
+            if self._weights_dirty:
+                args, auxs = self._module.get_params()
+                for name in self._arg_names:
+                    self._args[name][:] = args[name]
+                for name in self._aux_names:
+                    self._auxs[name][:] = auxs[name]
+                self._weights_dirty = False
+
+        def _set_weights(self, arg_params=None, auxs_params=None):
+            if self._module.binded:
+                self._module.set_params(self._args if arg_params is None else arg_params,
+                                        self._auxs if auxs_params is None else auxs_params,
+                                        allow_missing=True)
+                self._weights_dirty = arg_params is not None or auxs_params is not None
+            else:
+                if arg_params:
+                    for k in arg_params:
+                        self._args[k][:] = arg_params[k]
+                if auxs_params:
+                    for k in auxs_params:
+                        self._auxs[k][:] = auxs_params[k]
+                self._weights_dirty = False
+
+        def _update(self, updates):
+            for exe in self._module._curr_module._exec_group.execs:
+                outs = exe.output_dict
+                args = exe.arg_dict
+                for dst, src in updates.items():
+                    args[dst][:] = outs[src + '_output']
+
+        def _make_train_function(self):
+            def train_function(inputs):
+                data, label, _, data_shapes, label_shapes = self._adjust_module(inputs, 'train')
+
+                batch = mx.io.DataBatch(data=data, label=label, bucket_key='train',
+                                        provide_data=data_shapes, provide_label=label_shapes)
+                self._module.forward_backward(batch)
+                self._module.update()
+                self._update(self._train_updates)
+                self._weights_dirty = True
+                outs = self._module.get_outputs()[:self._ntrain]
+                return [x.asnumpy().mean() for x in outs]
+
+            self.train_function = train_function
+
+        def _make_test_function(self):
+            def test_function(inputs):
+                # although this function do testing we need the training symbol
+                data, label, _, data_shapes, label_shapes = self._adjust_module(inputs, 'test')
+
+                batch = mx.io.DataBatch(data=data, label=label, bucket_key='test',
+                                        provide_data=data_shapes, provide_label=label_shapes)
+                self._module.forward(batch, is_train=False)
+                if self._test_updates:
+                    self._update(self._test_updates)
+                    self._weights_dirty = True
+                outs = self._module.get_outputs()[:self._ntrain]
+                return [x.asnumpy().mean() for x in outs]
+
+            self.test_function = test_function
+
+        def _make_predict_function(self):
+            def predict_function(inputs):
+                data, label, _, data_shapes, label_shapes = self._adjust_module(inputs, 'pred')
+                batch = mx.io.DataBatch(data=data, label=label, bucket_key='pred',
+                                        provide_data=data_shapes, provide_label=label_shapes)
+                self._module.forward(batch, is_train=False)
+                if self._test_updates:
+                    self._update(self._test_updates)
+                    self._weights_dirty = True
+                outs = self._module.get_outputs()[:self._npred]
+                return [x.asnumpy() for x in outs]
+
+            self.predict_function = predict_function
+
+    return MXNetModel
